requirements.txt
requests
pandas
beautifulsoup4
vaderSentiment
python-dotenv
